Add files via upload
Daugman's algorithm is a widely used method for iris recognition, which is a biometric identification technique that uses the patterns in the iris of the eye for authentication and identification purposes. It was developed by John Daugman, a British computer scientist, and has been recognized as one of the most accurate and reliable methods for iris recognition.

The iris is the colored part of the eye that surrounds the pupil. It has a unique and intricate pattern of radial lines, furrows, and crypts that are stable over a person's lifetime. These patterns can be captured using specialized cameras and then processed to create a template that represents the individual's iris.

Daugman's algorithm consists of several key steps:

Image acquisition: An image of the iris is captured using an infrared camera or another imaging device that can capture detailed iris patterns. Infrared light is commonly used as it can penetrate the outer layers of the eye and capture the iris pattern more accurately.

Preprocessing: The captured image is preprocessed to enhance the iris pattern and remove any noise or artifacts. This may involve techniques such as normalization, image enhancement, and segmentation to isolate the iris region.